chore: Add integration test for todoapp with absolute path
While at it, made the formatting consistent. Also pulled out the shared
TEMPDIR & TEMPDIR2 into scoped variables. Sharing the same value between
two tests can lead to suble bugs. We want a new & tempdir for each test.

The need to project init & update rigth after cloning the todoapp makes
wonder if we shouldn't commit the cue.mod instead... Will open this
question in that repo instead. This does the job in this context. I have
confirmed locally that this test fails as expected with v0.2.23 & works
with this change. If you are wondering why not v0.2.24, this answers it:
https://github.com/Homebrew/homebrew-core/pull/105746

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>