AWS SAM package improvements (#2890)
* fix: implement switch - ciKey - to use tcp socket when running on gitlab ci

Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* fix: broken CI due to alpha/aws/cdk test

CI seems to be broken from a test on the aws.cdk package. It checks the python version with `python version` which doesn't seem to exist. Also, upgrades the node version as the cdk.#Image package relies on the latest version

Peer: gerhard

Signed-off-by: guillaume <guillaume.derouville@gmail.com>
Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* Uploads evaluated plan to DC when validation fails.

If plan validation fails, like checking action's inputs and outputs,
this enables the plan to be uploaded to DC if the user is logged in

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>
Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* ci: print cue lint feedback for unformatted files

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>
Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* Add README.md for the sam package

Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* Add code documentation and README for the sam package

Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* Fix linting errors and update Readme

Signed-off-by: Patrick Döring <munichbughunter@gmail.com>

* chore: Update new doc script to use ' instead of "

Some of our code editors keep changing these, this is will prevent some
of the churn.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>

* docs: Add AWS SAM page under the Use Cases section

This helps keep all content in the same place. It continues to be
available in GitHub, but also https://docs.dagger.io. The best part is
that we can use individual CUE & YAML files to keep the formatting valid
& consistent across all examples.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>

* fix: Use specific aws-sam-cli-build-image-provided

`latest` is a moving target and this might break if we don't pin it to a
specific version. Use the same #Image approach as we do in our main
packages, such as bash. FWIW:
https://github.com/dagger/dagger/discussions/2781#discussioncomment-3141568

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>

* fix: Add sam.#DefaultImage deprecation

We do not want to break this for (potential) existing users.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>

Co-authored-by: guillaume <guillaume.derouville@gmail.com>
Co-authored-by: Marcos Lilljedahl <marcosnils@gmail.com>
Co-authored-by: Gerhard Lazu <gerhard@lazu.co.uk>