ci: Disable GHA cache for Test Universe / Universe Tests
The last commit confirms that without the GitHub Actions cache, the
Experimental workflow passes. The last run was also quicker than
previous ones. While it is too early to tell if this is better after
just one run, this PR is one step closer from getting merged, so that
sounds like a step in the right direction.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>