Cache compilation of extensions.
This may seem like premature optimization, but without this every time
an extension is installed, every single **path** in its dependency DAG
will be walked, which scales very poorly to say the least. Even though
buildkit itself results in the builds being deduplicated, there was
already a noticeable slowdown when running the todoapp, which is still a
relatively simple dependency DAG.

This fixes the issue by using a map+mutex to cache calls to
RemoteSchema.Compile. It also uses a singleflight group to enable these
installs to run in parallel with one another.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>