Temporary back-compat with previous cloak dev -p flag.
playwithcloak stopped working, not sure but my best guess is that it was
using the `-p` flag to specify a port, but now `-p` is short for
`--project` and the full `--port` has to be used w/ cloak dev.

Until we have access to playwithcloak again, trying this temp fix to get
back-compat.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>