Add initial SDK support; remove Dockerfile boilerplate.
Now extensions are configured by specifying an SDK (currently ts, go and
dockerfile are supported). This is not yet the full "sdk-as-extension"
model as all these SDKs are just hardcoded LLB in the cloak server.
However this is a first step in that direction and squashes the
Dockerfile boilerplate in the meantime.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>