fix: Default to a rootless Buildkit container
Instead of using --privileged with the default buildkit container, use a
rootless one which makes it safer:
https://github.com/moby/buildkit/blob/master/docs/rootless.md#docker

One reason to *not* remove remove privileged is that our pipelines will
break, as well as our users'.

This is the first step towards attempting to start addressing
https://github.com/dagger/dagger/issues/151

While this does not solve the issue, it will inform us if this is a step
in the right direction and, more importantly, if this is still worth fixing.

We (+JF) are on a mission to close old issues!

Co-authored-by: Jean-François Farge <jf@dagger.io>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>