Fail auto-release if no changes since previous release (#3143)
We want to fail auto-releases on purpose if there have been no changes
since the previous release. While this can happen, it is rare and we
should take notice of it, hence the failure.

For more context, see this Discord thread:
https://discord.com/channels/707636530424053791/796905486145683506/1022141805468328027

My intention has been to update this to Dagger for a long time now. It
kept getting trumped by other priorities, and while now is the wrong
time to do it, every time I have to touch this YAML, I sigh and tell
myself: "one day...". Not today though.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Jeremy Adams <jeremy@dagger.io>