reorganize api schema and use go:embed
rather than using big ol' Go strings for Schema(), use go:embed so we
can start treating the schema definitions as the source of truth.

this means we'll need to start making concessions here and there until
the API is fully built out. as a concrete example, the git schema
returns a Filesystem for now until Directory is implemented and a TODO
has been added.

* renamed the files to .graphqls which is apparently more strictly
  appropriate.
* dropped .schema from code/test .go filenames now that the schema lives
  under schema/

Signed-off-by: Alex Suraci <alex@dagger.io>
Signed-off-by: Vikram Vaswani <vikram@dagger.io>