Fix type assertion when looking for input mounts.
The new generic wrapper resulted in go types of FSID to sometimes end up
as string (due to a marshal/unmarshal cycle). Unfortunately, there is
some fragile code which relied on that in the project code, specifically
for finding input mounts to include with extensions.

The root problem is that the code is too fragile and should be updated
to be more robust by using the graphql schema information rather than go
types. However, this issue is currently breaking the netlify extension
(results in the mount not being there) so just going with this fix right
now.

I've now noted this problem in the issue for improving this feature:
https://github.com/dagger/dagger/issues/3074#issuecomment-1262869710

Signed-off-by: Erik Sipsma <erik@sipsma.dev>