initial Container implementation
implements the beginnings of the new Container schema:

  container {
    from
    rootfs
    workdir
    withWorkdir
    variables
    withVariable
    exec
    exitCode
    stdout
    stderr
  }

also: tweak the handling of scratch directories so it can be re-used for
scratch containers. previously marshaling llb.Scratch() and then
unmarshaling it back into a st llb.State would cause an error when
calling st.Marshal. now we'll just set the definition to nil in the ID.

Signed-off-by: Alex Suraci <alex@dagger.io>