move schema implementations to core/schema/
core/ is now for domain models only, reducing its dependency footprint
and preventing import cycles.

this refactor is 95% just shifting code around, the only change is that
I added (*core.Container).WithSecretVariable, since that was implemented
inline in the schema and relied on (core.ContainerID.decode) which the
schema no longer has access to.

Signed-off-by: Alex Suraci <alex@dagger.io>