docs: Move files in version-specific dirs (#3319)
docs: Move files in version-specific dirs

Before this change, it was difficult to know which version of Dagger a
doc file was for. We had to look at the displayed_sidebar value inside
the doc file to know whether it was for Dagger v0.1.x or v0.2.x. 

After this change, most doc files are now in either v0.1 or v0.2 dirs.

Doc files which are not version specific are stored currently in one of
the following dirs:
- common
- unspecified
- use-cases

For more context, see https://github.com/dagger/dagger/issues/3318

Closes #3318

Signed-off-by: Vikram Vaswani <vikram@dagger.io>