Update all automation to use v0.2.x branch
Any automation which is not branch-specific (e.g.
add-issues-to-backlog.yml) is now in the `cloak` branch which will soon
be merged into main. See https://github.com/dagger/dagger/pull/3374 for
more details.

We remove dependabot.yml since this branch is going into maintenance
mode and we do not want to be dealing with dependency bumps.

The auto-release has now been disabled on this branch. It is now simpler
since we only expect to run in manually & very rarely.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>