Fix aws:cdk:python tests
Since we don't install a specific version of npm or node, we should not
be asserting the version in tests. We do it just for Python.

This test was broken for a long time, but because we are using a cache,
we didn't know about it. We should remove caching, for sure. That is a
separate PR, for now we want to get this merged and while fixing this
test is a tangent, it's small enough and worth taking.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>