refactor away from multi-return-value Decode()
This same refactor has already been applied to Container because one day
we needed to add another value to return. The Go linter complains
(justifiably) once you hit 4 non-error return values.

Another refactor I'm starting on will result in needing another value in
the File and Directory payloads, so this just does an initial structural
refactor in preparation of that.

Signed-off-by: Alex Suraci <alex@dagger.io>