Only run the lint workflow if **.go or **.md files change
We don't want to run this unnecessarily. We also don't want to run this
on tags. If we tagged something that was not linted correctly, something
is very wrong in our release workflow and should be addressed. FWIW:
https://github.com/dagger/dagger/blob/77eed15d356836981e4c0975920674069e315edf/.github/workflows/release-v0.3.x.yml#L87-L95

We want to run on larger runners so we learn quicker about potential
errors.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>