Only run lint & test when specific files change (#3444)
- Only run the lint workflow if **.go or **.md files change
- Only run the test workflow when Go-related files change
  + Also JS files in the sdk/nodejs dir

We don't want to run this unnecessarily. We also don't want to run this
on tags. If we tagged something that was not linted correctly, something
is very wrong in our release workflow and should be addressed. FWIW:
https://github.com/dagger/dagger/blob/77eed15d356836981e4c0975920674069e315edf/.github/workflows/release-v0.3.x.yml#L87-L95

Added a note to keep the node-version in sync with package.json. The
test will fail anyways (because of the engine constraint), but the
comment will hopefully make the fix more obvious.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>