docs: Add documentation for the CUE SDK and update the site navigation
After this change:
- A new 'docs/current' directory will be created for the current SDK-aware docs
- The latest Go and CUE docs will live in 'docs/current/sdk/[go|cue]'
- Doc UUIDs for the 'current' version will use numeric 6 chars
- All internal links for the 'current' version will use the new doc UUIDs
- All Go and CUE doc URLs in the 'current' version will be of the
  form '/sdk/[go|cue]/UUID/*'
- The sidebar will use a new structure
- The previous system for common files (files shared between versions) will no
  longer exist. A new system based on partials will replace it
- The 'current' documentation will appear at the root URL /
- The previous documentation will exist at the new URL /0.2 but no sidebar
  link in the 'current' version will exist to it

Keeping in mind that we have significantly more content for CUE than Go
and this creates an imbalance, this commit proposes a restructure of the
previous sidebar to make it single level/flat for both SDKs.
For the CUE SDK, this commit creates several separate landing pages
(reference docs, usage examples and guides) which contain the links
previously in the sidebar. This change should make the imbalance in content
slightly less visible.

Signed-off-by: Vikram Vaswani <vikram@dagger.io>