replace mounts rather than stacking them
"Stacking" in this commit means mounting over a path rather than
replacing it. On Linux this is allowed and unmounting the upper mount
will reveal the lower one.

Buildkit however currently assumes mounts have unique paths:

* GetMount returns the first one listed, rather than the last, which is
  what you'd want for stacked mounts as it's the mount that was actually
  visible to the process.
* Many APIs (e.g. Marshal) internally sort the mounts by the target
  name, implying that they don't expect collisions and marshaling may
  become nondeterministic without stable sorting.

So instead we'll just replace any mounts at the exact same target path.
It is no longer possible to unmount it to reveal the prior mount.

We could change Buildkit to not do this sorting and to have GetMount
return the upper mount, but for now this seems like the safest step
assuming this is not a common use case. We can always re-introduce
stacking if/when Buildkit supports it.

Signed-off-by: Alex Suraci <alex@dagger.io>