docs: Add Python SDK docs
This commit adds Python SDK documentation and a new set of nodes to the
navigation tree, It adds the Python SDK introduction, installation steps,
get started guide and additional guide. It adds the code for the guides
as individual code snippets. The additional guide provides a multi-build
example similar to the one used for the Go SDK.

Signed-off-by: Vikram Vaswani <vikram@dagger.io>