support dagger in dagger
Add the `ExperimentalPrivilegedNesting` flag to `Exec()` which sets a
valid `DAGGER_HOST` inside the container, allowing to communicate back
to the same engine instance.

In short, this means that it's possible to run a program that uses the
Go SDK within Dagger itself -- `dagger.Connect()` will automagically
work out of the box.

Signed-off-by: Andrea Luzzardi <aluzzardi@gmail.com>