feat: improve connect function to use a callback.
Previously, connect was initializing a new Cloak
Server and returned a ready to use GQL client.
The problem with that solution is that the user
has to explicitly manage the closing of the
server.

Now, connect takes a callback as parameter which
acts like a context where the user can execute
any query in it.
Finally, no matter what happens, connect will
close the server client.
This improvement simplify the code.

This commit also has few code style improvement
and remove LocalDir options from ConnectOpts
since it's not useful anymore.

Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>