ci: Revert use of sdk:node:test in test GHA workflow
Until we have DOCKER_HOST provisioning support in NodeJS, this does not
work. Let's keep it WIP so that the rest of the changes in this PR can
get merged & we can build on top of them (especially the publish part).

When the new provisioning support is added, and we have confirmed that
`./hack/make sdk:node:test` works locally, we can revert this.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>