sdk: python: Fix README.md example
Example had a wrong reference to the old `Result` wrapper.

Removed comments with `log_output` tip because it’s easy to miss and it’s covered in the getting started guide on the documentation site. Added progress status instead to keep terminal clean.

Kept the version in README.md as simple as possible. That’s just a carrot, the proper example is in the `examples` directory and in the docs.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>