ci: Extract Node.js SDK lint & test in a standalone workflow
We do not want to run all the tests for everything at once. This makes
the Node.js SDK tests run only when necessary.

This started with us realising that we were not running the tests when
some important files would change.

We should probably do the same for the other SDKs & engine too.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>