ci: cli distribution
This commit changes the tagging strategy of the engine to use the git
commit SHA instead of the registry hash. The main reason for this is to
avoid chicken and egg dependencies on some engine-dependent components (like
the CLI).

We're also moving the client-gen to its own cmd package since we're
removing the cloak command.

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>