Change shim to be an oci-runtime wrapper.
Now shim is no longer built on the fly, instead it's built into our
engine image. We then include it in dagger execs by configuring
buildkitd to use *it* as the oci runtime binary. It then updates the
container configuration to setup itself as the shim, which results in
the same code executing as previously.

This new approach has the short term benefit of stopping us from needing
to pull in golang images at runtime and build the shim on the fly, which
is a frequent user complaint.

In the longer term, the ability to hook into the container runtime
unlocks a ton of other interesting possibilities.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>