docs: Add deprecation note, update /install page
This commit adds a new deprecation notice to all documentation
pages under /0.1 and /0.2.

The content of the deprecation notice is stored as a partial.

This commit also creates a new page at the mutable /install
endpoint with links to the SDK installation pages. The
previous /install page is relocated to a new URL /1242/install.
Links are adjusted in older navigation bars.

Signed-off-by: Vikram Vaswani <vikram@dagger.io>