docs: Add API and CLI documentation
This commit adds documentation for the API and the CLI.
This includes reference documentation for both, installation
steps for the CLI, tutorials and guides for both, and an
introduction to the API Playground. This commit also updates
the general FAQ with entries related to the GraphQL API.

This commit updates the website navigation to include links
to the above documentation. This commit also adds a new
component that produces a button linking to ia Playground URL and
opening that URL in a separate tab/window.

Signed-off-by: Vikram Vaswani <vikram@dagger.io>