ci: Always run test & lint workflows (#3953)
This started as a follow-up to
https://github.com/dagger/dagger/pull/3897#issuecomment-1323917409

The TL;DR is that we want to keep as little logic in the GitHub Actions
workflows as possible, and keep as much as we can in Dagger (especially
the caching part).

The only other change that this introduces is running engine tests with
race detection in parallel. We want to optimise for the entire test
pipeline finishing as soon as possible, which means running everything
in parallel. Before this change, regular tests would need to pass first
before the tests with race detection would run.

Other changes:
- Run lint on all branches, not just main
- Small cosmetical improvements to lint workflow YAML
- Standardise all SDK publish workflows

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>