Update nested dagger to serve a single session.
This is partially needed just due to the updates around the new
_EXPERIMENTAL_* env vars interpreted by the SDKs. But it also ends up
being a pretty big simplification.

When dagger-in-dagger is enabled, the shim just opens a new session and
serves it over localhost, setting DAGGER_SESSION_* in the same way that
dagger run does.

There's no longer any need for mounting in a bin or the buildkitd socket
and no needed for any socket proxying of the session/buildkit api.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>