chore: Rename test-dagger-engine to dagger-engine.dev
Because it is a dev dagger-engine that was built from the current source code.

My first instinct was to rename this to dagger-engine-dev, similar to
dagger-engine-<SHA>, but there is garbage collection that uses the
dagger-engine- prefix, and I didn't want to refactor that in this
context. Yes, I was effectively side-stepping that problem since I
didn't want to be pulling too much change at once. That is due a
refactoring soon, so no point in digging into it now.

By the way, if you have an existing test-dagger-engine running, you will
need to remove it manually. You will also need to prune the volume. I
recommend:

    docker rm -f test-dagger-engine
    docker volume prune

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>