chore: Rename WithDevEngine to AdvertiseDevEngine
WithDevEngine implies that the wrapped container will use the dev
engine. This is misleading and results in surprising behaviour, as
described in https://github.com/dagger/dagger/pull/4300#issuecomment-1379391184

AdvertiseDevEngine is more accurate: the dev engine is being advertised
for any component that is interested in using it, but nothing is
enforced. As we discovered, some components will continue using the
engine in which they already run (this used to be the stable engine).
The link above goes into more detail - it includes screenshots &
diagrams that explain the state better than text alone.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>