Update SDK test instructions in releasing notes.
We've updated our publish workflows to automatically test SDKs
provisioning dagger directly on the host on both Linux and MacOS, so
instructions to do that manually were outdated.

However, we didn't specify that you should run through the get started
guide for each SDK to verify the publish worked, so that is added here.

Automating a run through of the get started guides seems like a good
candidate for automation as part of the publish workflow for each SDK.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>