ci: Collect traces & logs from pipeline runs and publish to Loki & Tempo  (#4340)
* dagger-graph: fix label escaping

Signed-off-by: Andrea Luzzardi <al@dagger.io>

* ci: add sub-pipelines

Signed-off-by: Andrea Luzzardi <al@dagger.io>

* ci: add lint/test meta targets

Signed-off-by: Andrea Luzzardi <al@dagger.io>

* otel-collector: send journal to tempo/loki

Signed-off-by: Andrea Luzzardi <al@dagger.io>

* ci: enable telemetry for engine test

Signed-off-by: Andrea Luzzardi <al@dagger.io>

* Replace GRAFANA_CLOUD_API_KEY secret with env variable
* Replace OTEL_EXPORTER_OTLP_HEADERS with env variable

No need for these to be secret since they can only write logs, traces & metrics. All repository forks should be able to submit the journal.log results.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Introduce low-cardinality labels for easier filtering

Rename labels to tags. These values are high-cardinality, we don't want
to be creating streams for every combination.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Replace all occurrences of label with tag

Uses tag when converting to JSON so that fields are prefixed with tag_ in Loki / Grafana.

Add run_url so that we can jump back to GitHub Actions.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Capture trace_id in each log line

So that we can correlate all logs that belong to a trace.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Link to the new "Dagger Overview" dashboard instead

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Add hostname to logs

This can be most helpful when comparing cattle with pets. Some pets are worth the hassle...

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Upgrade to latest stable go.dagger.io/dagger release

So that we no longer need the nightly.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Give runs a name that can be tracked long-term

We want to be able to see how these runs change over time. If we give them different IDs on every run, we will not be able to visualise the continuity easily.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Update metricsURL links

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Use runner name & arch in tags

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Andrea Luzzardi <al@dagger.io>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Gerhard Lazu <gerhard@dagger.io>