ci: Improve how we handle cached vertices (#4498)
./hack/make engine:dev and ./hack/make engine:test feeding data into the
same journal.log resulted in spans that didn't make sense. Some spans
would switch to a later group (i.e. engine:test), yet the started &
completed timestamps would clearly belong to a former group (i.e.
engine:dev).

When we did update the started & completed timestamps, we realised that
we were overwriting previous vertices, the uncached ones, with cached
timings, which also didn't make sense.

The easiest thing is to only track the first vertex, the uncached
version, and ignore the cached ones. This means that once a vertex is
solved in i.e. engine:dev, we don't show it again in engine:test, since
it's cached and less significant timing-wise (it rounds down to
0 micros).

In order to do this better, we would need to use more than the vertex
digest for the map. If we were to add the pipeline and/or the session
token, then we would be able to track the first cached vertex too. Since
it's not that much of a big win, it's OK to stop here for now, until we
have something better than a trace to visualise a pipeline run.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Andrea Luzzardi <al@dagger.io>

* Exclude .git & bin directories

Makes a big difference: 12MB vs 84MB. Maybe that is not the case when we
release - let's check before merging.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Andrea Luzzardi <al@dagger.io>