docs: Drop @latest from the package
In the case of Go, this will default to the latest published package
version, which is what we want. Versioned is unlikely to break in the
future, while `latest` is a moving target. In this context, moving
targets usually lead to unexpected breakages. I never regretted pinning
my dependencies. I consider it good advice at this point, but feel free
to drop this commit if you think otherwise.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>