remove explicit service starting/detaching
This was added so that the Git API could use services, but it's better
to express an explicit dependency even in that case.

Allowing an explicit Start makes code stateful and order-dependent. If
we decide that's OK after all, we can revert this commit, but for now it
seems better to continue down the path of having explicit dependencies.

Signed-off-by: Alex Suraci <alex@dagger.io>