support service dependencies for HTTP
this one's a little bit different; HTTP just returns a File directly so
there's no opportunity for a withServiceDependency API.

there isn't a really great reason for HTTP and Git to be different in
this regard. maybe Git should also just take it as an option? it's not
clear if it would ever need more than one.

Signed-off-by: Alex Suraci <alex@dagger.io>