go sdk: Wait on session child processes.
Without this the child proc became a zombie, which isn't good if lots of
clients are being made in a long-running process.

Additionally, this may fix an issue where it seems that the Go GC is
collecting the stdin pipe object for the child process before the
process has closed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>