Remove import dependency of CLI on containerd packages.
Code imported by the CLI was importing a constant from the new
remotecache package in dagger, but that created a transitive dependency
on some containerd packages which only build for Linux. This was causing
builds of the CLI for darwin to fail.

This works around that issue by reversing the import dependency so that
the remotecache package imports from `internal/engine` instead.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>