raise error when container has not been executed
Previously exitCode, stdout, and stderr would simply return null if a
container had not been executed yet, i.e. a container that's really just
an image. Worse, this would actually get turned into the zero-value for
each in the Go SDK (0, "", "") making it look like a successful exec.

Additionally, it was easy to accidentally use an un-exec'd container as
a service, e.g. by running an off-the-shelf redis or postgres image.
You'd get a weird error with an empty hostname.

Now in all cases we'll raise an error when trying to use a container
that has never exec'd.

Signed-off-by: Alex Suraci <alex@dagger.io>