Use singleton cache manager for experimental s3 cache pooling. (#4715)
This fixes a number of bottlenecks and problems in the previous cache
pooling implementation by switching to a new model where a single
manager in the engine process handles all import/export from/to the
remote target.

The previous approach was based on all the existing buildkit remote
cache implementations where the export and import code ran independently
for each client session. We already moved to a model where the cache was
configured server-side, but we were still using that same approach of
handling each session on its own.

This creates problems for pooling because each session to the engine
resulted in a new manifest being created. Even with pooling support this
created bottlenecks very quickly just due to the large number of
manifests.

It also didn't work because it forced us to just "stack" cache imports
where we'd import each manifest in the pool and then buildkit would
check each one in order every time it needed to check the cache at all.
This is as opposed to actually merging the cache manifests into one DAG.

---

The new approach creates a single s3 cache manager daemon when the
engine starts. This manager maintains in memory the current cache
manifest for the whole engine and then handles synchronizing that
between incoming export requests and new manifests it finds in the
remote cache pool from other engines.

When an cache export comes in from a client session, it now just merges
the cache manifest of that session to the daemon's engine-wide manifest,
which is just in memory metadata merging rather than any expensive I/O.
The manager will then asynchronously export the manifest to s3.

The manager also imports manifests from the remote pool by checking
under the pool's prefix and merging in cache metadata from there. It
does this once at engine startup and then periodically after that.

The overall performance benefits of this are:
1. Far fewer manifests accumulate in the pool, just one per engine
   instance (and the "name" of that engine can be re-used across time,
   you just can't have two engines with the same name running at the
   same time)
2. Clients don't block waiting for cache export to happen anymore. Once
   their build is done it all happens in the background.
3. Equivalent exports/imports of the same cached results will be deduplicated.
4. Rather than stacking cache manifests, they are now merged into one
   DAG, which should make querying much faster overall.

NOTE: at this point all of these performance benefits are still
theoretical and need to be verified in read world test cases in actual
clusters. However, the included integ tests verify that it all basically
works.

---

Part of this also involved moving entirely off of the existing buildkit s3
cache implementation code. As part of this I also found some low-hanging
optimizations such as using the s3 upload manager to upload layers whereas
the other implementation seemed to read each layer into memory fully and
then uploaded it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>