Container: add Run helper
This is a convenience API to replace this pattern:

```go
code, err := container.ExitCode(ctx)
if err != nil {
  return err
}
if code != 0 {
  return fmt.Errorf("exit code: %d", code)
}
```

Right now it's redundant because ExitCode returns an error when it
encounters a nonzero exit code, which is a bug (#3192). Anyone relying
on that behavior and ignoring the code return value should switch to
this new API instead.

Signed-off-by: Alex Suraci <alex@dagger.io>