Improve releasing instructions post v0.4.0
After going through the Engine + CLI v0.4.0 release, as well as all SDKs
& docs, these improvements made sense:

- last updated badge
- new high-level diagram
- convert 👉 action items to [x] checklist items
- optimise for top to bottom consumption

The only contentious item might be the removal of post-release checks,
i.e. https://github.com/dagger/dagger/pull/4461 . The manual checks are
so simple - checking the `dagger version` output - that if we messed
that up in our release process, we should add checks that would prevent
a release from going through if that was to happen. Let's discuss this
further in the PR.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>