fix color bleed in certain situations
e.g. ava.js output

an empty format should be treated like a reset, which is how it was
before. the Reset flag is to explicitly handle the case where a reset
followed by a foregroud/background sequence are combined, which won't be
an empty Format, and would unintentionally skip the reset.

Signed-off-by: Alex Suraci <alex@dagger.io>